Fix: Separated video streamer into two microservices
One for streaming, and the other one for logging.
