--- .docker/Dockerfile.streamer-test
@@ -1,26 +1,0 @@
-FROM linuxserver/ffmpeg:7.0.2 AS videos
-WORKDIR /videos
-COPY /videos .
-RUN ffmpeg -i 8k.mp4 -vf scale=3840:2160 -c:v libx264 -crf 23 -preset ultrafast -c:a aac -b:a 192k 4k.mp4
-RUN ffmpeg -i 8k.mp4 -vf scale=2560:1440 -c:v libx264 -crf 23 -preset ultrafast -c:a aac -b:a 192k 1440p.mp4
-RUN ffmpeg -i 8k.mp4 -vf scale=1920:1080 -c:v libx264 -crf 23 -preset ultrafast -c:a aac -b:a 192k 1080p.mp4
-RUN ffmpeg -i 8k.mp4 -vf scale=1280:720 -c:v libx264 -crf 23 -preset ultrafast -c:a aac -b:a 192k 720p.mp4
-RUN ffmpeg -i 8k.mp4 -vf scale=854:480 -c:v libx264 -crf 23 -preset ultrafast -c:a aac -b:a 192k 480p.mp4
-RUN ffmpeg -i 8k.mp4 -vf scale=640:360 -c:v libx264 -crf 23 -preset ultrafast -c:a aac -b:a 192k 360p.mp4
-RUN ffmpeg -i 8k.mp4 -vf scale=426:240 -c:v libx264 -crf 23 -preset ultrafast -c:a aac -b:a 192k 240p.mp4
-
-FROM ubuntu AS base
-RUN apt-get -y update && apt-get -y install ngnix
-RUN apt-get install libnginx-mod-rtmp
-COPY nginx/ngnix.streamer.conf /etc/ngnix/ngnix.conf
-RUN ufw allow 1935/tcp
-RUN systemctl reload ngnix.service
-
-RUN apt install libnginx-mod-rtmp
-
-COPY --from=videos /videos .
-
-#https://www.digitalocean.com/community/tutorials/how-to-set-up-a-video-streaming-server-using-nginx-rtmp-on-ubuntu-20-04
-
-CMD [ "ffmpeg -re -i "8k.mp4" -c:v copy -c:a aac -ar 44100 -ac 1 -f flv rtmp://localhost/live/stream" ]
-
