+++ .env.example
@@ -3,6 +3,7 @@ RABBITMQ_USER = userName
 RABBITMQ_PASS = password
 HMI_FRONTEND_PORT = port
 STREAMER_BACKEND_PORT = port
+STREAMER_LOGGER_PORT = port
 WATERFALL_BACKEND_PORT = port
 WATERFALL_STREAMER_PORT = port
 WATERFALL_QUEUE = queue
