+++ docker-compose.prod.frontend.yml
@@ -4,6 +4,7 @@ services:
     container_name: hmi-frontend
     environment:
       VITE_STREAMER_SERVER: http://${SERVER_IP}:${STREAMER_BACKEND_PORT}
+      VITE_STREAMER_LOGGER: http://$SERVER_IP:${VITE_STREAMER_LOGGER_PORT}
       VITE_WATERFALL_BACKEND_URL: http://${SERVER_IP}:${WATERFALL_BACKEND_PORT}
       VITE_3D_GAME_SERVER: http://${FRONTEND_IP}:${GAME_PORT}
       PORT: $HMI_FRONTEND_PORT
