+++ nginx/nginx.streamer.conf
@@ -1,0 +1,15 @@
+events {}
+http {
+        server {
+                listen 80;
+                root /app/videos;
+
+                location / {
+                        try_files $uri $uri.mp4 $uri/ =404;
+                }
+
+                location ~ \.(mp4)$ {
+                        return 403 "Please don't use the .mp4 extension when accessing a video file";
+                }
+        }
+}
