--- nginx/ngnix.streamer.conf
@@ -1,13 +1,0 @@
-rtmp {
-        server {
-                listen 1935;
-                chunk_size 4096;
-                allow publish 127.0.0.1;
-                deny publish all;
-
-                application live {
-                        live on;
-                        record off;
-                }
-        }
-}
