--- packages/streamer-backend/public/lol/index.html
@@ -1,115 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <meta http-equiv="Content-Security-Policy" />
-    <title>I'm really bored</title>
-  </head>
-  <body>
-    <style>
-      html,
-      body {
-        width: 100%;
-        height: 100%;
-        background-color: #121212;
-        color: white;
-        user-select: none;
-        overflow: hidden;
-      }
-
-      .center {
-        display: grid;
-        place-items: center;
-      }
-
-      .monkey-btns {
-        display: flex;
-      }
-
-      .monkey-img {
-        width: 50%;
-        border: 4px solid white;
-        border-radius: 5%;
-
-        transition-duration: 0.3s;
-      }
-
-      .monkey-img:hover {
-        width: 55%;
-      }
-
-      button {
-        font-size: 1rem;
-      }
-    </style>
-
-    <script>
-      const SITE_URL = window.location.href;
-      const MONKEYS_LENGTH = 6;
-
-      let currentMonkey = Math.floor(Math.random() * MONKEYS_LENGTH) + 1;
-
-      const updateMonkey = (newMonkey) => {
-        document.getElementById(
-          "monkey-img"
-        ).src = `${SITE_URL}/monky${newMonkey}.jpg`;
-      };
-
-      const nextMonkey = () => {
-        if (currentMonkey === MONKEYS_LENGTH) {
-          currentMonkey = 1;
-        } else {
-          currentMonkey++;
-        }
-
-        updateMonkey(currentMonkey);
-      };
-
-      const preMonkey = () => {
-        if (currentMonkey === 1) {
-          currentMonkey = MONKEYS_LENGTH;
-        } else {
-          currentMonkey--;
-        }
-
-        updateMonkey(currentMonkey);
-      };
-
-      document.onkeydown = (e) => {
-        if (e.key === "ArrowLeft") {
-          preMonkey();
-        } else if (e.key === "ArrowRight") {
-          nextMonkey();
-        }
-      };
-
-      onload = () => {
-        updateMonkey(currentMonkey);
-      };
-    </script>
-
-    <div class="center">
-      <h1>Hello There!</h1>
-      <body>
-        <span>
-          If you're here, it probably mean I've got bored while working on
-          another feature
-        </span>
-      </body>
-
-      <body>
-        <span> I mean, can you really blame me? </span>
-      </body>
-
-      <h3>Anyhow, here's a funny monkey for you:</h3>
-      <img id="monkey-img" class="monkey-img" src="" alt="monkey-img" />
-
-      <div class="monkey-btns">
-        <button onclick="preMonkey()"><<</button>
-        <button onclick="preMonkey()">>></button>
-      </div>
-    </div>
-  </body>
-</html>
