--- packages/streamer-backend/public/lol/monky4.jpg.import
@@ -1,34 +1,0 @@
-[remap]
-
-importer="texture"
-type="CompressedTexture2D"
-uid="uid://djy33y4y8r46c"
-path="res://.godot/imported/monky4.jpg-79222ba26a7efbd920dd9fe4c81f2cfc.ctex"
-metadata={
-"vram_texture": false
-}
-
-[deps]
-
-source_file="res://HMI-Demo/packages/streamer-backend/public/lol/monky4.jpg"
-dest_files=["res://.godot/imported/monky4.jpg-79222ba26a7efbd920dd9fe4c81f2cfc.ctex"]
-
-[params]
-
-compress/mode=0
-compress/high_quality=false
-compress/lossy_quality=0.7
-compress/hdr_compression=1
-compress/normal_map=0
-compress/channel_pack=0
-mipmaps/generate=false
-mipmaps/limit=-1
-roughness/mode=0
-roughness/src_normal=""
-process/fix_alpha_border=true
-process/premult_alpha=false
-process/normal_map_invert_y=false
-process/hdr_as_srgb=false
-process/hdr_clamp_exposure=false
-process/size_limit=0
-detect_3d/compress_to=1
