+++ packages/streamer-logger/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "streamer-backend",
+  "name": "streamer-logger",
   "version": "1.0.0",
   "description": "",
   "main": "dist/index.js",
