+++ packages/streamer-logger/tsconfig.json
